Graphics.cpp: Added a static_cast to UINT32 to remove size difference with 64 bit ulonglong ( size_t ). Vec2.cpp: Chanced the way the self normalize function works handling cases when vector length is 0.f. Font class: After changing to Win 8.1 SDK, compiler no longer liked the way I friended the DWrite class and Font classes.  Added an operator-> to DWrite for access to the underlying IDWriteFactory pointer.

+++ Engine/Graphics.cpp
@@ -58,7 +58,7 @@ Graphics::Graphics( HWNDKey& key )
 void Graphics::DrawString( const std::wstring & Str, const RectF & Rect, const Font & F, Color C )
 {	
 	pBrush2d->SetColor( ConvertColor( C ) );
-	pContext2d->DrawText( Str.c_str(), Str.length(), F, Rect, pBrush2d.Get() );
+	pContext2d->DrawText( Str.c_str(), static_cast< UINT32 >( Str.length() ), F, Rect, pBrush2d.Get() );
 }
 
 Graphics::~Graphics()
